Added staggered entrance animation for portfolio chips
