+++ src/pages/dashboards/playground-ai.astro
@@ -2,12 +2,13 @@
 import PageLayout from "@/layouts/PageLayout.astro";
 import { PlaygroundAiComponent } from "@/components/portfolio/dashboards/playground-ai";
 import ContainerSection from "@/components/ContainerSection.astro";
-import HeaderPortfolio from "@/components/HeaderPortfolio.astro";
+import HeaderPortfolio from "@/components/portfolio/common/header-portfolio";
 ---
 
 <PageLayout meta={{ title: "Playground AI" }}>
 	<ContainerSection>
 		<HeaderPortfolio
+			client:visible
 			title={"Playground AI"}
 			description={"A responsive user interface for interacting with AI models or chatbots, intuitive and user-friendly. (in progress)"}
 			chipsList={["React", "Typescript", "Tailwind CSS", "Shadcn UI", "Radix UI"]}
