+++ src/pages/random-person.astro
@@ -2,12 +2,13 @@
 import PageLayout from "@/layouts/PageLayout.astro";
 import ContainerSection from "@/components/ContainerSection.astro";
 import RandomPersonComponent from "../components/portfolio/random-person";
-import HeaderPortfolio from "@/components/HeaderPortfolio.astro";
+import HeaderPortfolio from "@/components/portfolio/common/header-portfolio";
 ---
 
 <PageLayout meta={{ title: "Random Person" }}>
 	<ContainerSection maxWidth="max-w-2xl">
 		<HeaderPortfolio
+			client:visible
 			title={"Random Person"}
 			description={"A random profile generator for testing purposes, complete with contact details and a photograph."}
 			chipsList={["React", "Typescript", "react-icons", "Rest API", "Tailwind CSS"]}
