+++ src/pages/svg-anime-universe.astro
@@ -1,7 +1,7 @@
 ---
 import SvgAnimeUniverseRoot from "@/components/portfolio/svg-anime-universe";
 import PageLayout from "@/layouts/PageLayout.astro";
-import HeaderPortfolio from "@/components/HeaderPortfolio.astro";
+import HeaderPortfolio from "@/components/portfolio/common/header-portfolio";
 ---
 
 <PageLayout
@@ -10,6 +10,7 @@ import HeaderPortfolio from "@/components/HeaderPortfolio.astro";
 >
 	<div class="w-full overflow-hidden pb-20 pt-32">
 		<HeaderPortfolio
+			client:visible
 			title={"SVG Anime Universe"}
 			description={"Homepage view of an anime service with advanced SVG animation of moving circles."}
 			chipsList={["React", "Typescript", "SVG API", "CSS"]}
