+++ src/pages/task-manager.astro
@@ -2,12 +2,13 @@
 import PageLayout from "@/layouts/PageLayout.astro";
 import ContainerSection from "@/components/ContainerSection.astro";
 import TaskRoot from "@/components/portfolio/task-manager";
-import HeaderPortfolio from "@/components/HeaderPortfolio.astro";
+import HeaderPortfolio from "@/components/portfolio/common/header-portfolio";
 ---
 
 <PageLayout meta={{ title: "Task Manager" }}>
 	<ContainerSection maxWidth="max-w-2xl">
 		<HeaderPortfolio
+			client:visible
 			title={"Task Manager"}
 			description={"Task management tool that simplifies daily organization and enhances productivity."}
 			chipsList={["React", "Typescript", "react-redux", "@reduxjs/toolkit", "Context API"]}
